Add boatDetailTab data refreshing
On the Details tab of boatDetailTab, data was not refreshing in response
to edits made through the boatSearchResults lightning-datatable. To
correct this, I added type and payload properties to messages in the
BOATMC message channel, allowing distinction between types of messages.
Then, I added a new 'refresh' message, which alerts the tab that it
should clear the cache and get its data from the db. This also happens
any time a new boat is selected from the gallery.